Add support for SSL when testing for redirects
I ran into issues with using this cookbook in an air-gapped environment
where the Nexus binaries were hosted over SSL on a web server with a
certificate issued by an internal Certificate Authority.

This patch ensures that the pre-download check, which is intended to
follow redirect responses, provides the `net/http` library with the
proper port number to issue the request against and enables SSL if the
download URL uses the `https` scheme.