Ensure all vmoptions have a new line
The init script reads each line of nexus.vmoptions and appends it
to the start command, but for it to read the line, it must end with
a new line. If there is no new line, the option is ignored, which
if using only join("\n") is the case of the last option.

This commit ensures all options have a new line, so none are ignored.