Fix delete action on resources role, task, user & cleanup_policy
In role, task & cleanup_policy the enclosing "converge_if_changed" was
preventing the resource to actually delete the object when it exists.

In user & role the missing 'not_if' guard prevent the idempotency.

In the delete script for cleanup_policy, a typo has been introduced so
the groovy code was incorrect.