Allow linking routing rule to repository
We cannot use attributes to associated a routing rule, so
we must use a separate param, like online.

The groovy scripts need to be updated, to retrieve the routing
rule ID when linked it, and its name when fetching the repo infos.

Adding a depencies recipe, so we can create resources that can be
reused in resources test. (Lost too much time on it already, so maybe
at some point I'll look into custom ordering the generated resources
to not have another recipe)