api: only generate json for json content
So we can reuse the request method helper, only
generate JSON when we set application/json content
type.

For scripts, switch to application/json as this is what
is expected when we have more than just a string. Even
though we were sending text/plain, we were generating
JSON (because Hash or Array), and Nexus3 was happy with
it.