do not check service for fedora
There is a systemd bug affecting fedora (29, 30, ...),
which makes `systemctl show` fail and so prevent us from
checking the service being enabled / running.

We check that the port is running and make some API calls,
so even if we do not check via systemd, we known the
service is running.

Skip os.linux? which includes fedora and some other os we
do not test. Centos is part of  os.redhat, so not affected.
https://docs.chef.io/inspec/resources/os/