Support license installation
Via system property path so we do not require a restart on
new install or upgrade as the license will be directly loaded.
cf: https://help.sonatype.com/en/installing-and-updating-licenses.html#installing-or-updating-a-license-using-a-system-property

But also via the API directly for when the license needs
to be updated as the system property path does not care
about the license file being updated if a license is
already installed.